+++ app/src/components/map/Tools/EditTools.tsx
@@ -1,7 +1,6 @@
-import { Button, IconButton, Paper, Popover } from '@material-ui/core';
 import { useContext, useEffect, useRef, useState } from 'react';
 import { useLeafletContext } from '@react-leaflet/core';
-import { useMapEvent, GeoJSON } from 'react-leaflet';
+import { useMapEvent } from 'react-leaflet';
 import * as turf from '@turf/turf';
 import L from 'leaflet';
 import React from 'react';
@@ -9,138 +8,7 @@ import single from '../Icons/square.png';
 import multi from '../Icons/trim.png';
 import { async } from 'q';
 import { Capacitor } from '@capacitor/core';
-import { LayersControlProvider } from '../LayerPicker/layerControlContext';
-
-const POSITION_CLASSES = {
-  bottomleft: 'leaflet-bottom leaflet-left',
-  bottomright: 'leaflet-bottom leaflet-right',
-  topleft: 'leaflet-top leaflet-left',
-  topright: 'leaflet-top leaflet-right'
-};
-const interactiveGeometryStyle = () => {
-  return {
-    color: '#3388ff',
-    weight: 5,
-    opacity: 0.5,
-    stroke: true,
-    clickable: true,
-    fill: false
-  };
-};
-
-const MobilePolylineDrawButton = ({ convertLineStringToPoly, setGeometry, context }) => {
-  const positionClass = POSITION_CLASSES.topleft;
-  const [anchorEl, setAnchorEl] = React.useState<HTMLButtonElement | null>(null);
-  const [drawMode, setDrawMode] = useState(false);
-  const [locArray, setLocArray] = useState([]);
-  const divRef = useRef();
-  const open = Boolean(anchorEl);
-  const id = open ? 'polylineDraw' : undefined;
-
-  const [geoToConvert, setGeoToConvert] = useState(null);
-
-  useEffect(() => {
-    if (locArray.length > 1) {
-      setGeoToConvert({
-        type: 'Feature',
-        geometry: {
-          coordinates: locArray,
-          type: 'LineString'
-        },
-        properties: {}
-      });
-    }
-  }, [locArray]);
-
-  useEffect(() => {
-    if (divRef?.current) L.DomEvent.disableClickPropagation(divRef?.current);
-  });
-
-  useMapEvent('click', (e) => {
-    const loc = e.latlng;
-    if (drawMode) {
-      setLocArray([...locArray, [loc.lng, loc.lat]]);
-    }
-  });
-
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const deleteLastPoint = () => {
-    if (locArray.length > 1) {
-      const tempArr = [];
-      for (var i = 0; i < locArray.length - 1; i++) {
-        tempArr.push(locArray[i]);
-      }
-      setLocArray(tempArr);
-    }
-    if (locArray.length === 1) {
-      setLocArray([]);
-    }
-  };
-
-  const drawModeChange = () => {
-    setDrawMode(!drawMode);
-    handleClose();
-    if (drawMode) {
-      const poly = convertLineStringToPoly(geoToConvert);
-      setGeometry([poly]);
-      setGeoToConvert(null);
-      setLocArray([]);
-    }
-    if (!drawMode) {
-      context.clearLayers();
-      setGeoToConvert(null);
-    }
-  };
-
-  return (
-    <LayersControlProvider value={null}>
-      <div ref={divRef} className={positionClass}>
-        <Paper>
-          <IconButton
-            className="leaflet-control leaflet-bar"
-            style={{
-              borderRadius: 5,
-              height: 46,
-              width: 46,
-              position: 'absolute',
-              marginTop: 76
-            }}
-            onClick={handleClick}></IconButton>
-        </Paper>
-        <Popover
-          id={id}
-          open={open}
-          anchorEl={anchorEl}
-          disableEnforceFocus={true}
-          disableRestoreFocus={true}
-          anchorOrigin={{
-            vertical: 'top',
-            horizontal: 'right'
-          }}
-          transformOrigin={{
-            vertical: 'top',
-            horizontal: 'left'
-          }}>
-          <div style={{ display: 'flex', flexFlow: 'row nowrap', backgroundColor: '#a0a098', borderRadius: 5 }}>
-            <Button onClick={drawModeChange}>{drawMode ? <>Finish</> : <>Start</>}</Button>
-            <Button onClick={deleteLastPoint}>Delete last point</Button>
-            <Button onClick={handleClose}>Cancel</Button>
-          </div>
-        </Popover>
-      </div>
-      {
-        geoToConvert && <GeoJSON key={Math.random()} data={geoToConvert} style={interactiveGeometryStyle} /> //NOSONAR
-      }
-    </LayersControlProvider>
-  );
-};
+import { MobileDrawCancel, MobilePolylineDrawButton } from './Helpers/MobileDrawBtns';
 
 const circleORmarker = (feature, latLng, markerStyle) => {
   if (feature.properties.radius) {
@@ -186,6 +54,7 @@ const formulateTable = (feature) => {
 const EditTools = (props: any) => {
   // This should get the 'FeatureGroup' connected to the tools
   const [multiMode, setMultiMode] = useState(false);
+  const [openCancel, setOpenCancel] = useState(false);
   /* Removed toggling multimode for now:
   const toggleMode = () => {
     setMultiMode(!multiMode);
@@ -266,11 +135,12 @@ const EditTools = (props: any) => {
     if (!multiMode) {
       (context.layerContainer as any).clearLayers();
     }
+    setOpenCancel(true);
   });
   useMapEvent('draw:deleted' as any, () => {
     props.geometryState.setGeometry([]);
   });
-  useMapEvent('draw:deletestop' as any, () => onDeleteStop);
+  useMapEvent('draw:deletestop' as any, onDeleteStop);
   useMapEvent('draw:edited' as any, onEditStop);
 
   const convertLineStringToPoly = (aGeo: any) => {
@@ -448,6 +318,7 @@ const EditTools = (props: any) => {
           context={context.layerContainer}
         />
       )}
+      {openCancel && Capacitor.getPlatform() === 'ios' && <MobileDrawCancel setOpenCancel={setOpenCancel} />}
       {/*<IconButton
         //ref={divRef}
         className={themeContext.themeType ? toolClass.toolBtnDark : toolClass.toolBtnLight}
